new try

+++ hashmap.c
@@ -128,7 +128,7 @@ Pair * nextMap(HashMap * map) {
   if(map->current==-1){
     return NULL; 
   }
-  for(int i = map->current+1; i < map->capacity; i++){
+  for(long i = map->current+1; i < map->capacity; i++){
     if (*(map->buckets+i) != NULL && (*(map->buckets+i))->key != NULL) { 
     map->current = i;
     return *(map->buckets+i);
